[AMD] Fix DeepSeek-OCR-2 multimodal token mismatch for local crops
images_crop contains normalized pixel values (floats in [-1, 1]) but
was cast to torch.long, truncating most values to 0. When has_local_crops
was not propagated through MultimodalDataItem, the fallback
torch.sum(patches).item() != 0 evaluated to False for certain images,
causing the model to skip local crop embeddings while the processor
already emitted placeholder tokens for them. This produced a fatal
RuntimeError: Insufficient multimodal embedding length.

Fix both sides:
- Cast images_crop to target_dtype (same as pixel_values) instead of
  torch.long so the fallback check works correctly.
- Add has_local_crops and has_images to ATTR_NAME_TO_MODALITY so they
  propagate through the multimodal processor pipeline, eliminating the
  need for the fallback entirely.

+++ python/sglang/srt/models/deepseek_ocr.py
@@ -1686,7 +1686,7 @@ class DeepseekOCRForCausalLM(nn.Module):
 
         images_crop = (
             torch.stack([item.images_crop for item in mm_items], dim=0)
-            .type(torch.long)
+            .type(target_dtype)
             .to(device=pixel_values.device)
         )
         images_spatial_crop = (
